[Platinum II] Title: 가장 가까운 두 점, Time: 104 ms, Memory: 5124 KB -BaekjoonHub

+++ 백준/Platinum/2261. 가장 가까운 두 점/가장 가까운 두 점.cc
@@ -13,14 +13,14 @@ int main()
 {
     cin >> N;
     //x,y 순서
-    vector<pair<int,int>> pointsVec;
+    vector<pair<long long, long long>> pointsVec;
 
     //y,x 순서
     set<pair<int,int>> activeSet;
 
     while(N--)
     {
-        int x, y;
+        long long x, y;
         cin >> x >> y;
         pointsVec.push_back({x,y});
     }
